requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly==5.24.1
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
reportlab>=4.1.0
kaleido==0.2.1
openpyxl>=3.1.0
psycopg2-binary>=2.9.0